bookBack/requirements.txt
Django~=2.2.4
djangorestframework-jwt~=1.11.0
djangorestframework~=3.12.4
PyJWT~=1.7.1
pip~=20.0.2
wheel~=0.34.2
pytz~=2021.1
sqlparse~=0.4.1
setuptools~=41.2.0